Parallelise fuzzer
Most of the fuzzer overhead is waiting for act-fuzz, a fairly hefty
sequential program, to do its thing.  With this in mind, I've
rearchitected the multi-file fuzzer to divide the problem into N chunks,
where N is the number of CPUs.

I've seen the time for a 100-pass fuzz run drop by a factor of 3 on a
4-core CPU, so I feel as if this is probably worth it.  It may, however,
be worth considering the number of chunks later on.