Switch chunking to 1-per-subject + optional cap
After running a few back of napkin experiments with the chunking factor
set at various values between CPU count and input count, it seems that
more is always better, and so I've set it to default to input count.

I've kept the chunking mechanism around for now as a tunable 'cap' on
the number of chunks, but I have a feeling I'll eventually just bin it.

Ah well, there are worse ways to waste an afternoon I'm sure.