Shufti around parallelism
This adds a few new worker count endpoints (meaning that the config will need a few more `workers = BLAH` overrides),
swaps around the lifter parallelism so that it's done as a parallelised run on subjects with a sequential run on
arches, and does a few other things adjacent to parallelism.

There are still a few issues with parallelism that need working out (dangling worker counts, worker defaults,
compiler parallelisation, forwarding things properly from rmach to mach), but this is mostly trying to fix an
aggravated parallelisation file exhaustion problem.