Gratuitously rearrange plan litmus
Both original-litmus and fuzz-litmus now expand out into subrecords with
statsets.
This lets us add stat tracking to the fuzzer, which will become useful
as we inline the litmus patcher.

I've also added a lot of subject construction boilerplate, which may need
reworking as we go on, but tidies things up a little.

Not properly tested outside of unit tests yet, so it may fall over.