Somewhat overhaul analysis
We'll hopefully have proper plan extraction stuff in place soon, so
having a plan analysis format that is somewhere between human readable
and machine readable seems a bit ridiculous.  This nudges it into something
that looks vaguely like markdown, which should make it a bit easier to read.

We're also adding options to the analysis writer, which should get exposed
in the -analyse tool soon.