Add plan compression/decompression and deindentation
The idea at the moment is that 'invoke' uses deindented but uncompressed
plans (but this might change to compressed ones eventually, will need to
benchmark); archival uses indented but compressed plans (because they might
be human-read in future); and everything else continues as normal.

Reading from file now does a little dance over the first two bytes of the
file to see if there's a gzip magic number.  There is no such transparency
for general reads, but there is a new ReadMagic function that implements
the magic number detection for seekable readers.

Tests now check that read/write round trips work with all flag combinations.