Shuffle around compiler struct names
compiler.Compiler -> compiler.Configuration
compiler.Config -> compiler.Compiler

Confusing?  Perhaps, but the idea is that the first one is now a combination
of a compiler record with a 'configuration' of specifically how it is to be
invoked.

This commit might be a bit messy, as it's been pulled from a bigger changelist.