Check for nil before closing various remote handles
This fixes a nil-dereference error that occurs in the invoker,
where we try to close a runner after an error.
Sometimes, the error arrives from trying to construct a connection as part
of the runner, and the result is that the runner is in an incomplete state
with one of its closeable members being nil.

This fix might mask genuine nil pointer issues, but it's easier to think about
than trying to do anything more specific.