Add special matching logic for filtered statuses
When classifying flags, anything that is filtered gets only the 'filtered' status,
and no other.

All of the special casing pertains to matching against statuses rather than flags,
so I've split the flag.Matches function to produce a high-level and special case
handling flag.MatchesStatus function, and removed the FlagOk flag as it was a little
misleading.