Dash: track compiler count in border title
This is a fairly large change to effect a small quality of life
improvement (making it easy to eyeball how much I need to multiply the
cycle counts for to get compiler counts, at least until I get around to
making the stats board report that directly!  but also making it easy
to eyeball if compilers are missing).  That said, it sets precedent for
how we can do similar things in future, and also gave me an excuse to
migrate some of the dash code out into a separate struct as I've been
slowly doing all along.