Clean up director use of observers
- Get rid of loggers wherever they exist; if I need the information
  they were sending, I'll add observer calls instead.
- Remove the responsibility for closing observers out of the director
  and put it back in the app level...
- ...with a new directorobs.Obs struct for combining the 'usual
  suspect' loggers; the logger observer has moved here and the director
  observer package has collapsed into the director package.
- Move plan observation to the director observer; the current observers
  ignore plan messages, but I'll fix this eventually.

The idea is partly that I want to separate concerns a bit better (by
having the flow of control happen only in one direction where possible),
pave the way for having the director/director-instance work reused in
a single-shot context.

I also want to, tomorrow maybe, add some more file observers.  In
hindsight this'll probably need to be an extension of the Logger.