Lay groundwork for coverage testbed maker
I was going to make a shell script to do the coverage testbed, but
thought it would perhaps be more maintainable, testable, and
later-expandable-to-do-more-automated-stuff than a shell script.  Maybe
this was a horrendous mistake - only time will tell.

The maker doesn't do anything yet, but I've added a stub application, an
initial config file, and various other things.

The config file uses a new TOML library that I've been planning to
roll out across the tester at some point, but never got round to doing;
maybe this will be sufficient impetus.