Cov: use worker pool, make some things deterministic
Seeds for inputs are now just the number of the input in the profile.
Input file selections are randomised but using the same seed (0).

We now use a worker pool (default 10 workers) for orchestrating
profiles, but in practice the throughput seems to be in lockstep, and
I'm not entirely sure why.  Probably because there are only two threads
submitting jobs...?