Try to sort out the instance channels mess
Made a generic 'fan in channels' implementation, and rerouted the logger
to use it.  If all goes well, I might consider either abstracting out the
common legwork of the fan-in-director-observer, or re-architecting to make
it central.

This got started a few weeks ago and derailed by article writing,
so I'm not sure if I've done what I originally set out to do.