Lay more groundwork for standalone herd/litmus
I've added legs to the two that actually run herd and litmus in standalone mode,
but they're inaccessible and untested.

I'm building towards having the recipe contain instructions for the run bit of
the machine node, and having standalone produce an instruction just to the tune of
'cat the output file'.