Move interpreter to separate package
To allow for use in the runner phase.  (Eventually, I'd still like to get rid of
the separate phases and have some sort of elaborate worker pool that handles
concurrency constraining for runs, maybe with lock/unlock or semaphore
instructions, but we'll see.)

compiler.Driver has moved to interpreter, for now, but I'm not entirely convinced
it belongs there.