Flail around even more on interpreter
I've changed tack to make output types a single enumeration on the recipe,
with the idea being that if the recipe is text then the machine runner'll just
skip compilation and go to catting the input files.

I actually need to *implement* that, of course, and I'm nowhere close.

I've been rearranging things to clean up bits of the compiler area, prep for
more flexible recipes, etc.  Hm.