Decouple forwarding logic from logger
The director logger is now a 'ForwardHandler', which specifies behaviour
over forwarded cycle observations, and plugs into a 'ForwardObserver'
that handles the actual forwarding.

Per my TODO in directorobs.go, I'm imagining that eventually the
forwarding logic will become implicit (with ForwardHandlers just
becoming director Observers).  The dashboard currently doesn't adhere
to forwarding, and re-engineering it to do so would be some effort, so
currently it's a separate abstraction layer.