Expand architecture information in litmus tests
This horrifically expansive commit mostly hinges around adding architecture fields,
architecture querying from files, cleaner construction, and other gubbins to Litmus
tests.

A lot of the third bit of this has rippled through and caused changes everywhere a
litmus test gets built.  Ah well.

Tests all pass, but there's a lot of new code and not a lot of it is covered yet.