Add rudimentary grace period support
This is important to make the backend runner able to support partial traces; the
default CommandContext setup always sigkills, and we need to sigterm things like
rmem to get them to dump a partial result.

Seems to work; could do with some testing, but how?, and refactoring, but how?.

This isn't yet supported in the machine node, mainly because that doesn't quite
support the service.Runner idiom yet.

See https://github.com/golang/go/issues/22757.