Fix accidental breakage of litmus patcher
This happened because your esteemed code monkey made a dependency implicit in the
flow of the litmusifier, and then refactored the litmusifier so that it calculated
its arguments to pass to Litmus BEFORE calculating which arguments to enable.  Oops.

Refactored so that the fixset is passed around explicitly, which is a bit ugly but
also makes the dependency explicit.