Drop backend specs in backend methods
This requires backends to include any information from the spec in themselves;
currently, the only thing we use from the spec (if anything) is the RunInfo,
and so the backend resolver now maintains a list of functions that map from
RunInfo overrides to specifications.

This change cleans up the API quite significantly, and opens up some more
possibilities.  One possible issue may be racing on backends, as we now often
construct the backend before sending it to multiple threads.