Add first draft of auto-selector
I've decided to split implementation of this into two things: the director
boilerplate that wraps kills in observations and mutant selection in whatever
I need to implement to get that to work, and this, which just handles the
mutant selection.

This isn't fully tested yet; tests only cover the cause where killing the
mutant advances the auto-selector.  I'll get back to that after adding mutant
interpolation, I think.