Move state accounting outside state runners
Runner implementations need to provide a Stage() stage.Stage implementation,
and the various support code for running stages now expect and use it, instead
of the stage accounting being done inside the implementations.

This paves the way to turn the stage queue in the director into a simple
[]Runner, and then to add and remove things from the Runner as needed.

Changes plan version.