Implement rudimentary experiment deadlines
We now have a somewhat more precise notion of experiment start time (ie the
time when loops begin executing), and we report this as a prepare message.

At the same time, we also now have a timeout system which can set a duration
after that start time (plus a little bit of CPU time in between, so not 100%
accurate) at which the context for the loops
will be killed.

This allows us to set experiments with a time limit, say 24 hours, and walk
away in the knowledge that the system will gracefully stop the experiment at
that point.

Fixes #124.

No tests yet, as I'm not entirely sure how they'd work.