Thread timespans through plan/analysis/stats
This is a fairly invasive battery of changes that try to unify how c4t stores
and does calculations on timespans, including logging them for hits and kills
in mutation testing.

This has quite heavily broken some tests in ways that were hard to salvage, so
some of the tests are not as good at covering as they used to be, especially
pretty-printer regressions.  To be fixed later on.

Changes plan version.