Implement c4t-config print-compilers
The main reason I'm doing this is so I can TDD the implementation of machine
architectures and architecture inheritance later.

This involves also implementing a rudimentary abstraction for tabulating
things (to CSV or to a tabwriter).  I'll likely start porting the existing CSV
code to this as time goes by.