Start implementing database schema
We don't actually commit any analysis to the database yet, primarily as I
haven't worked out exactly what the schema should be.
But this is the first step.

We now have a `c4t-mkdb` command for executing the DDL to start a sqlite
database running.  Eventually, this might not be needed, as the database might
be auto-created.  It's useful for testing, anyway.